backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aio-pika
amqp
anyio
asgiref
attrs
awscli-local
billiard
boto3
botocore
celery
certifi
cffi
charset-normalizer
click-didyoumean
click-plugins
click-repl
click
coreapi
coreschema
django-debug-toolbar
django-filter
django-rest-swagger
django
djangorestframework-simplejwt
djangorestframework
fastapi
h11
idna
iniconfig
itypes
jinja2
jmespath
kombu
localstack-client
logger
markupsafe
multidict
openapi-codec
packaging
pamqp
pika
pluggy
prompt-toolkit
psycopg2-binary
py
pycparser
pydantic
pyjwt
pyparsing
pytest
python-dateutil
pytz
rabbitmq
requests
s3transfer
simplejson
six
sniffio
sqlparse
starlette
tomli
typing-extensions
uritemplate
urllib3
uuid
uvicorn
vine
wcwidth
yarl
djangoviz
pytest-django
pytest-cov
pytest-factoryboy
django-extensions
